requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
Django==5.0.7
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
itemadapter==0.9.0
itemloaders==1.3.1
jiter==0.5.0
jmespath==1.0.1
lxml==5.2.2
openai==1.43.0
packaging==24.1
parsel==1.9.1
pillow==10.4.0
Protego==0.3.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
python-telegram-bot==21.4
queuelib==1.7.0
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
scrapy-djangoitem==1.1.1
service-identity==24.1.0
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tldextract==5.1.2
tomli==2.0.1
tqdm==4.66.5
Twisted==24.3.0
typing_extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
webencodings==0.5.1
zope.interface==6.4.post2
